+++ src/IP.php
@@ -3,8 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\IP;
 
-use Innmind\IP\Exception\DomainException;
-
 /**
  * @psalm-immutable
  */
@@ -27,7 +25,7 @@ abstract class IP
     final public static function v4(string $address): IPv4
     {
         if (!\filter_var($address, \FILTER_VALIDATE_IP, \FILTER_FLAG_IPV4)) {
-            throw new DomainException($address);
+            throw new \DomainException($address);
         }
 
         /** @psalm-suppress ArgumentTypeCoercion $address cannot be empty here */
@@ -40,7 +38,7 @@ abstract class IP
     final public static function v6(string $address): IPv6
     {
         if (!\filter_var($address, \FILTER_VALIDATE_IP, \FILTER_FLAG_IPV6)) {
-            throw new DomainException($address);
+            throw new \DomainException($address);
         }
 
         /** @psalm-suppress ArgumentTypeCoercion $address cannot be empty here */