+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 
 - Require PHP `8.4`
 
+### Removed
+
+- `Innmind\IP\Exception\Exception`
+- `Innmind\IP\Exception\DomainException`
+
 ## 3.2.0 - 2023-09-16
 
 ### Added